% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{advdate}
\alias{advdate}
\title{Advance dates by week}
\usage{
advdate(startdate, week)
}
\arguments{
\item{startdate}{The first date in the series (e.g. first day of class).}

\item{week}{The number of weeks into the series.
Week 1 is the first date in the series.}
}
\value{
A vector of dates in the format "dd mmm" in parentheses.
}
\description{
This function takes a start date, e.g. of a course,
and advances it by a number of weeks.
This function is used in the syllabus template.
}
\examples{
advdate("2021-06-23", 1:3)
}
