% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_profile.R
\name{bookdown_profile}
\alias{bookdown_profile}
\title{Manage iheiddown profile}
\usage{
bookdown_profile()
}
\value{
Opens or creates a bookdown profile YAML file for the current thesis.
}
\description{
This RStudio Addin opens up the ~/_bookdown.yml file for
users to edit. If the file doesn't exists, it will create one.
}
