% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\alias{preview_start}
\alias{preview_stop}
\title{Previewing rmarkdown documents}
\usage{
preview_start(file_path, ...)

preview_stop(which = NULL)
}
\arguments{
\item{file_path}{The input Rmd file path (if missing and in RStudio, the
current active document is used).}

\item{...}{Additional arguments passed to
\code{rmarkdown::\link[rmarkdown]{render}()}.}

\item{which}{An integer vector of the server IDs;
by default, IDs of all existing servers in the current R session obtained
from \code{daemon_list()}, i.e., all daemon servers will be stopped by default.}
}
\value{
Starts serving the current files to RStudio's Viewer pane.

Stops serving the current files to RStudio's Viewer pane.
}
\description{
Use the \pkg{servr} package to continuously serve and reload
the Rmd document. Run \code{preview_start()} function while viewing the Rmd
panel of your presentation. To stop the server, run \code{preview_stop()}.
}
\references{
xaringan
}
