% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_pdf.R
\name{thesis_pdf}
\alias{thesis_pdf}
\title{Creates an R Markdown PDF Thesis document}
\usage{
thesis_pdf(input = ".", ...)
}
\arguments{
\item{input}{Rmd input file that will be processed by the function.}

\item{...}{Instructions passed to the function from a correctly specified
.Rmd document}
}
\value{
A modified \code{pdf_document} based on the IHEID Thesis LaTeX
template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the IHEID Thesis LaTeX template file.
}
\examples{
\dontrun{
knit: iheiddown::thesis_pdf
}
}
